+++ index.js
@@ -45,10 +45,17 @@ module.exports = {
 
           const response = await self.apos.http.post(url, options);
           if (!response.success) {
+            self.logInfo('hcaptcha-invalid-token', {
+              ip: req.ip,
+              data: response
+            });
             throw self.apos.error('invalid', req.t('AposHcaptcha:invalidToken'));
           }
+          self.logInfo('hcaptcha-complete', {
+            ip: req.ip
+          });
         } catch (error) {
-          self.apos.util.error(error);
+          self.apos.util.error('hCaptcha error', error);
           throw self.apos.error('error', req.t('AposHcaptcha:captchaErr'));
         }
       }