+++ test/test.js
@@ -159,6 +159,12 @@ describe('Forms module', function () {
     let page = await apos.http.get('/', { jar });
     assert.ok(page.match(/logged out/), 'page contains logged out in body');
 
+    // intecept the logger
+    let savedArgs = [];
+    apos.login.logInfo = (...args) => {
+      savedArgs = args;
+    };
+
     await apos.http.post(
       '/api/v1/@apostrophecms/login/login',
       {
@@ -175,7 +181,56 @@ describe('Forms module', function () {
       }
     );
 
+    assert.equal(savedArgs[0], 'hcaptcha-complete');
+    assert(savedArgs[1].ip);
+
     page = await apos.http.get('/', { jar });
     assert.ok(page.match(/logged in/), 'page contains logged in in body');
+  });
+
+  it('should log bad token request', async function () {
+    const mary = getUserConfig();
+
+    const jar = apos.http.jar();
+
+    // establish session
+    const page = await apos.http.get('/', { jar });
+    assert.ok(page.match(/logged out/), 'page contains logged out in body');
+
+    // intecept the logger
+    let savedArgs = [];
+    apos.login.logInfo = (...args) => {
+      savedArgs = args;
+    };
+
+    try {
+      await apos.http.post(
+        '/api/v1/@apostrophecms/login/login',
+        {
+          method: 'POST',
+          body: {
+            username: mary.username,
+            password: mary.pw,
+            session: true,
+            requirements: {
+              AposHcaptcha: 'bad-token'
+            }
+          },
+          jar
+        }
+      );
+    } catch (error) {
+      //
+    }
+
+    assert.equal(savedArgs[0], 'hcaptcha-invalid-token');
+    assert(savedArgs[1].ip);
+    delete savedArgs[1].ip;
+    assert.deepEqual(savedArgs[1], {
+      data: {
+        success: false,
+        'error-codes': [ 'invalid-input-response' ]
+      }
+    });
   });
 });